+++ lib/ckb/types/output.rb
@@ -3,7 +3,7 @@
 module CKB
   module Types
     class Output
-      attr_accessor :data, :lock, :type, :out_point
+      attr_accessor :lock, :type, :out_point, :data
       attr_reader :capacity
 
       # @param capacity [String]
@@ -11,11 +11,11 @@ module CKB
       # @param lock [CKB::Types::Script]
       # @param type [CKB::Types::Script | nil]
       # @param out_point [CKB::Types::OutPoint | nil]
-      def initialize(capacity:, data: "0x", lock:, type: nil, out_point: nil)
+      def initialize(capacity:, lock:, type: nil, out_point: nil, data: "0x")
         @capacity = capacity.to_s
-        @data = data
         @lock = lock
         @type = type
+        @data = data
         @out_point = out_point
       end
 
@@ -24,6 +24,7 @@ module CKB
       end
 
       def calculate_bytesize
+        raise "Don't know data" if @data.nil?
         bytesize = 8 + Utils.hex_to_bin(@data).bytesize + @lock.calculate_bytesize
         bytesize += @type.calculate_bytesize if @type
         bytesize
@@ -33,13 +34,12 @@ module CKB
         Utils.byte_to_shannon(calculate_bytesize)
       end
 
-      def to_h(data = true)
+      def to_h
         hash = {
           capacity: @capacity,
           lock: @lock.to_h,
           type: @type&.to_h
         }
-        hash[:data] = @data if data
         hash[:out_point] = @out_point if @out_point
         hash
       end
@@ -51,7 +51,6 @@ module CKB
         out_point = OutPoint.from_h(hash[:out_point]) if hash[:out_point]
         new(
           capacity: hash[:capacity],
-          data: hash[:data],
           lock: Script.from_h(hash[:lock]),
           type: type,
           out_point: out_point