+++ lib/ckb/types/script.rb
@@ -7,7 +7,7 @@ module CKB
 
       # @param code_hash [String]
       # @param args [String[]]
-      def initialize(code_hash:, args:, hash_type: "Data")
+      def initialize(code_hash:, args:, hash_type: "data")
         @code_hash = code_hash
         @args = args
         @hash_type = hash_type
@@ -38,30 +38,17 @@ module CKB
         )
       end
 
-      def to_hash
-        blake2b = CKB::Blake2b.new
-        blake2b << Utils.hex_to_bin(@code_hash) if @code_hash
-        blake2b << case @hash_type
-                   when "Data"
-                     "\x0"
-                   when "Type"
-                     "\x1"
-                   else
-                     raise "Invalid hash type!"
-                   end
-        args = @args || []
-        args.each do |arg|
-          blake2b << Utils.hex_to_bin(arg)
-        end
-        blake2b.hexdigest
+      def to_hash(api)
+        api.compute_script_hash(self)
       end
 
-      def self.generate_lock(target_pubkey_blake160, system_script_code_hash)
+      def self.generate_lock(target_pubkey_blake160, secp_cell_type_hash, hash_type = "type")
         new(
-          code_hash: system_script_code_hash,
+          code_hash: secp_cell_type_hash,
           args: [
             target_pubkey_blake160
-          ]
+          ],
+          hash_type: hash_type
         )
       end
     end